+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 1.1.0
 * Adds comparator for all classes including Objectable.
+* Adds ability to update attributes
 
 ## Version 1.0.0
 * Allows creation of objects with raw data (i.e. no fancy conversion).